+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - root_pid.py accepts graceful termination timeout argument
 - added option --utc to log messages with UTC timestamp instead of local time
 - added option --no-security-update for offline creation of images
+- added option --reduce-size to create an reduced image without docs and so on
 
 ### Changed
 - refactored root_pid.py to depend only on python3-minimal